requirements.txt
flask>=2.0.0
selenium>=4.0.0
pyyaml>=6.0